+++ changelog.md
@@ -1,2 +1,6 @@
+# 1.1.0
+- Bongo Cat now also bongos when you jump using the keyboard
+- Added more hats and extra decorations
+
 # 1.0.0
 - Release